webhooks/discord/requirements.txt
psutil 
gputil 
requests
python-dotenv
discord.py